nexthop: fix disappearing addresses and routes after expiration
After a fixed period of time (20 minutes), grout will mark non-static
nexthops that were resolved earlier (with ARP or NDP) as STALE and will
try to probe them to refresh their mac addresses.

After 6 probes have been sent (3 unicast using the previously known mac
address of the nexthop, and 3 broadcast/multicast) without any reply
from the nexthop, it is marked as FAILED.

After 1 minute in the FAILED state, it is "freed" using the callback
specified in the nexthop_ops per family (at the moment, only IPv4 and
IPv6).

The current callbacks are rib4_cleanup and rib6_cleanup which are very
aggressive. They effectively delete everything related to the nexthop.
Including any local address, routes, etc.

Replace these callbacks with simpler ones which only delete /32 and /128
routes and decrement the reference counter on the nexthop. If the
nexthop still has routes referencing it (for example, it is used as
a gateway for a route), scrub its flags and mac address fields so that
it can reused later.

Fixes: fe3e408851ce ("route: rework route cleanup routine")
Signed-off-by: Robin Jarry <robin@jarry.cc>