smoke: add nexthop ageing test
Since the nexthop ageing parameters can now be tuned, add a basic test
that configures a bunch of addresses, check that nexthops are resolved
and that they are refreshed automatically after their expired lifetime.

Also ensure that if the peers do not reply to ARP requests, the nexthops
are properly destroyed without breaking the interface addresses
configuration.

Signed-off-by: Robin Jarry <robin@jarry.cc>