address: always allow deleting addresses
If a local address nexthop has routes referencing it (other than the
connected route created when the address is configured), it should be
allowed to delete the address in any case.

Remove the check for the ref_count value. rib4_cleanup and rib6_cleanup
will take care of removing any routes that are referencing the nexthop.

Signed-off-by: Robin Jarry <robin@jarry.cc>