worker: keep threads running all the time
Keep one worker thread per CPU in the datapath affinity mask. Only
destroy threads when removing CPUs from that mask. Now that threads are
waiting for a pthread condition, this will not consume any cycles on
these CPUs and will allow the kernel to change the C-state / P-state of
the cores accordingly.

Moreover, since the threads are now always started, moving rxqs around
should be faster.

Signed-off-by: Robin Jarry <robin@jarry.cc>