infra: send iface down notification before deleting iface
To support FRR, interfaces created or deleted in Grout must be synchronized
with FRR. FRR listens for interface notifications from Grout to keep its
internal state consistent.

When FRR receives an IFACE_EVENT_PRE_REMOVE notification, the dplane plugin
in FRR attempts to remove the corresponding interface from memory. However,
Zebra fails to delete it if the interface is still marked as "up".

In the Linux kernel, when an interface is deleted, a netlink message is
first sent to indicate that the interface is down. Grout should follow the
same behavior.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>