Revert "api: allow api access with net_admin apability"
This reverts commit 4a685c7da75bc5c7b642878216e69fd5f9580499.

The capability library operates per-thread, while SO_PEERCRED returns the
PID of the process, which corresponds to the TID of the main thread.

When a client connects to Grout, SO_PEERCRED provides the PID of the
connecting process—effectively the main thread’s TID. If the connecting
thread (not the main thread) holds CAP_NET_ADMIN, cap_get_pid will report
that the capability is not set, since it checks capabilities for the main
thread only. As a result, the connection is rejected, even though it should
be allowed.

Since there is no way to obtain the peer TID through getsockopt, this issue
cannot be resolved cleanly. Therefore, this commit is being reverted.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>