port: allow attaching already probed devices
At startup, we pass -a 0000:00:00.0 to rte_eal_init() to disable
automatic PCI probing. However, this only affects the PCI bus devices.
Other buses (e.g. FSMLC) are not affected by this argument and are
always probed in rte_eal_init().

When configuring a port with such a device, we get an error from the API
saying that the port is already attached to another interface:

	grout# add interface port dpni2 devargs fslmc:dpni.2
	error: File exists

Instead of iterating over the DPDK probed devices, look at grout ports
to determine whether a port matching this devargs value is already
configured.

If we cannot find one, check if the device has already been probed and
only probe it if it was not.

Signed-off-by: Robin Jarry <robin@jarry.cc>