meson: generate grout.pc - pkg-config
A pc/pkg-config file is needed in order to provide helpers for
integration. For instance, it can be used when FRRouting is crosscompiled
in order to avoid an unsafe path such as /usr/include/grout.

Note: On buildroot, when a local clone is used for development,
then there is an intermediate build directory that does not have a .git
metadata folder, so by default for such case, Buildroot will set the
version to "custom".
See buildroot/package/pkg-generic.mk:
 ifneq ($$($(2)_OVERRIDE_SRCDIR),)
 $(2)_VERSION = custom
 endif

Signed-off-by: Vincent Jardin <vjardin@free.fr>