ip: add stateless dnat44 support
Allow configuring stateless "destination NAT" addresses. This is
a convoluted term to represent the following:

                          client
                |                        ^
                | destination            |  source changed
                |    8.8.8.8             |   to 8.8.8.8
                v                        |
            +----------------------------------+
            |       iface p0    8.8.8.1/24     |
            |                                  |
            |              grout               |
            |                                  |
            |       iface p1    10.16.0.1/24   |
            +----------------------------------+
                 |                       |
          destination changed            |
             to 10.16.0.99               |
                 |                       |
                 v                       |
                    service 10.16.0.99

This enables grout for exposing virtual/floating IP addresses on one
interface without actually owning these addresses. And internally
translating the packet addresses to forward packets to another host
located behind another interface.

DNAT44 is always bi-directional and is configured from the "external" or
"public" standpoint. Packets received on the "public" interface with
a specific IP address will have their destination address changed to an
internal IP. Packets transmitted from the public interface that have
a source IP address matching the internal IP will have their source IP
changed to the "public" IP.

Internally, this feature is implemented using nexthops since the nexthop
themselves are required to support ARP and normal communication, we only
store the address to substitute on packets in a private field. Special
flags are added to nexthops and interfaces to make this work.

Only the external->internal direction is processed in a separate node.
The internal->external direction needs to be done in the middle of the
ip_output node (before sending packets to either eth_output or
ipip_output).

Signed-off-by: Robin Jarry <robin@jarry.cc>