modules: use topological sort with dependencies
Change the concept of init_prio and fini_prio to a single depends_on
field. This new field takes a module name that must be equal to another
module. When the field is left NULL, the modules will be initialized
first and finalized last (in no particular order). Other modules will be
initialized after (and finalized before) the modules they depend on.

The topological sort algorithm was inspired from multiple sources and
adapted for grout module objects.

Link: https://mohammad-imran.medium.com/understanding-topological-sorting-with-kahns-algo-8af5a588dd0e
Link: https://www.techiedelight.com/kahn-topological-sort-algorithm/
Signed-off-by: Robin Jarry <robin@jarry.cc>