main-loop: expose rcu object
Create an RCU object global to the application and initialize it with
RTE_MAX_LCORE reader threads. Report quiescent state every turn of main
loop.

Expose the rte_rcu_qsbr structure pointer for any subsystem that would
like to make use of it.

It will be leveraged in the next commits to ensure the datapath workers
do not hold a reference on objects when they are freed/unallocated by
the control plane thread.

Link: http://doc.dpdk.org/guides/prog_guide/rcu_lib.html
Signed-off-by: Robin Jarry <robin@jarry.cc>