nexthop: split type field in type and family
Store the address family in a separate field. Use this field only for
anything related to the address itself. The nexthop type is only used
for graph steering in ip{,6}_output.

Nexthop type now default to the GR_NH_T_L3 type which does not have
specific handling in the datapath.

Update the CLI code to display the type and address family. Update all
code accordingly.

Signed-off-by: Robin Jarry <robin@jarry.cc>