nexthop: introduce nexthop id
In the next commits we will need to identify specific nexthop that have
been created to reference them as target for routes.

Add an arbitrary unsigned 32 bit integer to nexthop objects. The value
is completely ignored by grout. Its only purpose is for users to
designate a specific nexthop. The special value 0 is considered as
"unset".

Update the config test to create and delete nexthop with IDs.

Signed-off-by: Robin Jarry <robin@jarry.cc>