nexthop: add route origin field
Include a new origin field in nexthops objects exchanged in the API.
Allow users to set an arbitrary value. This is mostly intended for
routing daemons to identify if a nexthop was learned from BGP, OSPF, etc.
Or if it was configured manually by the user.

Display the origin of nexthop in the CLI. For now, all nexthop explicitly
added by the CLI are GR_NH_ORIGIN_USER. Other API clients can specify
any suitable value.

GR_NH_ORIGIN_INTERNAL(255) is reserved for dynamically added nexthop
when learning new next hops after receiving ARP requests or replies.
Hide these nexthops on purpose when replying to a GR_NEXTHOP_LIST
request to avoid confusion.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>