ip,ip6: drop packets larger than output interface mtu
Currently, we don't check if the packets transmitted will fit in the
configured MTU of the output port.

Before proceeding further, drop the packets if they are larger than the
configured MTU.

Some day, we will need to support fragmentation for IPv4 and sending an
ICMP6 "packet too big" error for IPv6.

Signed-off-by: Robin Jarry <robin@jarry.cc>