nexthop: allow modifying existing nexthops
When exist_ok is set to true in a GR_NH_ADD request, do not ignore the
fields from the request when the nexthop already exists. Make sure to
override all fields with the provided values.

If the address family or IP address has changed, make sure to delete the
old /32 or /128 route before re-adding it with the updated values.

Trigger a GR_EVENT_NEXTHOP_UPDATE with the updated values.

It is now possible to do this:

  + grcli add nexthop 1.2.3.4 iface p0 id 42
  > nh new: iface 1 vrf 0 1.2.3.4 00:00:00:00:00:00
  > route add: 1.2.3.4/32 via 1.2.3.4 origin link

  + grcli add nexthop f00:ba4::666 iface p1 id 42
  > route del: 1.2.3.4/32 via 1.2.3.4 origin link
  > route add: f00:ba4::666/128 via f00:ba4::666 origin link
  > nh update: iface 2 vrf 0 f00:ba4::666 00:00:00:00:00:00

Signed-off-by: Robin Jarry <robin@jarry.cc>