treewide: replace include guards with pragma once
An alternative to include guards using `#ifndef`, `#define` and `#endif`
is `#pragma once`. This non-standard but commonly supported directive
among C and C++ compilers, it also requires less code and does not
require the definition of a variable with risk of name conflicts. Also,
the variable definitions often deviate after file renaming and/or move.

Bonus, the use of `#pragma once` may improve compilation speed since it
is a higher-level mechanism. The compiler itself can compare filenames
or inodes without having to invoke the C preprocessor to resolve
`#ifndef` and `#endif`.

Signed-off-by: Robin Jarry <robin@jarry.cc>