frr: add srv6 localsid sync from zebra to grout
This commit adds synchronization of SRv6 LocalSIDs from zebra to Grout.

We cannot create SRv6 nexthops, because they are of type NEXTHOP_IFINDEX
and Grout does not yet support nexthops without a gateway. Therefore, for
SRv6 we bypass the nexthop-ID API and create only the LocalSID when the
grout dplane plugin processes a GR_IP6_ROUTE_ADD with a srv6 nexthop.

Note: USD is always enabled on Grout, whereas on FRR/Linux it is disabled
by default and must be enabled manually. Until this is harmonized, zebra
logs a debug message if USD is not enabled.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>