infra/nexthop: derive vrf_id from iface in nh_add
In the Grout nh_add API, an interface is mandatory and already bound to a
VRF. Supplying a separate vrf_id is redundant and allows mismatched
configuration. Ignore the caller‑provided vrf_id and always take the VRF
from the interface.

Motivation: FRR has a ZAPI bug [1] where route installs create nexthops
with VRF index 0 instead of the route’s VRF. This doesn't affect the
kernel, as nexthops are not scoped by VRF. But it affects Grout, which
creates nexthops within the VRF of the interface. Passing a mismatched
vrf_id can result in a nexthop being incorrectly installed in VRF 0 while
the interface belongs to another VRF. Deriving the VRF from the iface in
nh_add prevents this mismatch.

Note: vrf_id was always set to zero. But in next commit, we are going to
support vrf with FRR and Grout. As said, dplane_ctx_get_vrf(ctx) always
returns 0.

[1] https://github.com/FRRouting/frr/blob/frr-10.4.0-rc1/zebra/zapi_msg.c#L2229
Signed-off-by: Maxime Leroy <maxime@leroys.fr>