icmp: fix usage of rte_pktmbuf_mtod_offset
Usage of a negative offset for rte_pktmbuf_mtod_offset is not only
ugly, but also doesn't guarantee that the header is actually valid.

Introduce a new helper to extend an existing mbuf data, so we can
add a new metadata to the existing priv.
This allows the icmp priv data to include the IP info.

Fixes: 0a399004 ("ip: add ping and traceroute commands")
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>