infra: don't call loopback_output from ip(6)_output
gr-loopX is currently used to send local packets in exception paths and
to route some traffic to the kernel. We plan to reuse gr-loopX as a VRF
interface to support route leaking between VRFs via FIB lookup
delegation.

To enable this, we register gr-loopX with
ip_output_register_interface_type(GR_IFACE_TYPE_LOOPBACK, vrf_output), so
packets can be injected into a target VRF. Since only one callback can be
registered per interface type, we must remove the current loopback_output
handler.

To route traffic into the kernel, a TAP interface can be created—similar
to how veth pairs are used to connect namespaces. This makes TAP devices
a more appropriate solution than gr-loopX for connecting GROUT to the
kernel or to other routers.

Note: the DPDK TAP driver currently breaks when the TAP interface is
moved to a Linux namespace, as it tries to issue ioctl() calls from the
wrong context. This needs to be fixed.

gr-loopX is now only used to send local packets to the control plane.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>