infra: fix error when loopback tun interface cannot be created
When running grout as !root or when running it in a network namespace,
the creation of the loopback tun network device may fail with obscure
errors:

  ERR: GROUT: iface_loopback_init: ioctl Operation not permitted

or:

  ERR: GROUT: iface_loopback_init: ioctl Device or resource busy

This should not prevent the creation of interfaces. It will only hinder
the linux communication (and maybe some cross VRF functionality). In
such cases, an error in the log will be enough to identify the issue.

Change vrf_incref to not return any error. Replace the aggressive ABORT
in vrf_decref with a WARNING log.

Fixes: ebdb8babc741 ("infra: avoid event API for VRF create/delete")
Signed-off-by: Robin Jarry <robin@jarry.cc>