infra: transform loopback netdev from tun to tap
In order to forward non IP frames to the kernel, a full ethernet frame
is required.
Remove the direct injection of packets to ip(6)_input nodes, and rely on
the common nodes to direct the packets to the proper destination.

loopback_ouput now behaves similarly to eth_output, prepending an
ethernet header.

On gr-loop creation, ignore the ifup message to not assign an ipv6
local address to have the same behavior as before.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>