infra: transform loopback netdev from tun to tap
In order to forward non IP frames to the kernel, a full ethernet frame
is required.
Remove the direct injection of packets to ip(6)_input nodes, and rely on
the common nodes to direct the packets to the proper destination.

loopback_ouput now behaves similarly to eth_output, prepending an
ethernet header.

Following kernel behavior, do not add an ip6 link local to the vrf
iface.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>