ip/ip6: fix display nexthop
With the ability to set a nexthop with an ID, we now support
a IPv4 gateway for a IPv6 network, and the other way around.

Fix the command 'show ip(6) route' to print the address with
the proper format.

grout# show ip route
VRF  DESTINATION       NEXT_HOP       ORIGIN  ID  NEXT_HOP_VRF
0    16.0.1.0/24       fc00:210::2    user    11  0

Fixes: a1e4087b ("all: add support for gateway-less nexthops")
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>