nexthop: cleanup on iface deletion
With the ability to insert a nexthop in the rib without being tied to a
specific route, we were lacking the proper cleanup on interface deletion.
While the nexthop was deleted properly, it wasn't removed from the RIB/RIB6.

Call the ops->del associated to AF_UNSPEC to remove the nexthop from all RIBs.

Test to verify nexthop/rib4/rib6 are properly cleaned on
interface deletion in config_test.sh

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>