nexthop: avoid nuking routing tables after 6 arp/ndp failures
When 3 unicast ARP/NDP and 3 {broad,multi}cast probes have failed to get
a response, the nexthop is marked as "failed" and is "deleted" after
lifetime_unreachable_sec (60 seconds by default).

This deletion is rather aggressive and completely removes all routes
that point to this failed nexthop. This may have unwanted side effects
when running for a long time in production.

Keep the nexthop alive and leave all routing tables alone.

Signed-off-by: Robin Jarry <robin@jarry.cc>