vec: add debug checks with magic marker
When compiling without NDEBUG (i.e. in debug mode), add an extra field
to the __gr_vec_hdr structure and ensure it always contains a magic
value. Every time a gr_vec_* macro is called, ensure the pointer
references a valid gr_vec header. If not, abort with an explicit error.

Signed-off-by: Robin Jarry <robin@jarry.cc>