github: test deb and rpm packages build in pull requests
We often have DEB,RPM build failures after merging PRs. This is awkward
and can be avoided if we test these in every pull request.

Move the deb and rpm jobs into a reusable package.yml workflow that
fires on all pull requests as well as on workflow_call. Invoke the
packaging jobs in the publish workflow with upload=true so that the
generated packages can be aggregated.

Signed-off-by: Robin Jarry <robin@jarry.cc>