nexthop: remove implicit /32 or /128 route
When creating a nexthop with the API, an implicit /32 or /128 route is
automatically added. This route is useless and was only added when the
API for creating nexthops was added to avoid leaving nexthops with
ref_count=0.

Remove that and simply increment the reference counter. If users want to
reference the nexthop they created in routes, they can do so afterwards.

Signed-off-by: Robin Jarry <robin@jarry.cc>