smoke: fix cross vrf tests configuration
These tests accidentally work because the gateway routes are resolved
first by pinging them explicitly.

In VRF 2, there is no route configured to reach 16.1.0.0/16. When
16.1.0.1 is received in VRF 1, it resolves to nexthop ID 2 (172.16.1.2)
which is configured in VRF 2. If 172.16.1.2's mac address has not been
resolved yet, the packet is sent to the control plane for it to send
an ARP request.

Unfortunately, at this point, mbuf_data()->iface has been changed to the
interface in VRF 2 and rib4_lookup() is performed in VRF 2 which returns
NULL and the packet is dropped.

Configure a route to reach 16.1.0.0/16 in VRF 2 as well and remove the
initial pings on 172.16.0.1 and 172.16.1.1 to ensure everything works as
expected.

Signed-off-by: Robin Jarry <robin@jarry.cc>