deb,rpm: enable frr plugin
Enable the frr option in order to clone and build FRR and build our
zebra dplane plugin against it.

Register a static version in the dependencies to avoid any ABI issues.

Generate a new container image with the upstream RPM packages for
frr-10.4.1 and our zebra dplane plugin installed in the correct
location. Publish the container to quay.io/grout/frr with an extra label
containing the frr version.

Unfortunately, the UBI9 image does not contain the required dependencies
for frr. It notably misses c-ares. Generate the container image based on
centos9 stream and start from scratch for the final image (decreases the
final image size from 300M to 130M).

Also base the quay.io/grout/grout image on scratch to save space.

Signed-off-by: Robin Jarry <robin@jarry.cc>