id-pool: make it safe for datapath multithread use
rte_bitmap is not thread safe. Also, it does not support random access.
Use atomic operations to ensure multiple writers and multiple readers
safety.

The implementation is inspired from npf portmap. It contains an
additional gr_id_pool_get_random() function that will be used by dynamic
SNAT in the next commit.

To save cycles when the id pool is half or almost full, represent it
with a two-level tree.

We still store the actual IDs in a flat bitmap of uint64_t but we add
another array that has a zoomed out version of the flat bit map. Each
bit in that zoomed out array represents one slab in the flat bitmap. If
a bit is set to 1, it means that the underlying bitmap has at least one
available bit. Use this to determine which slab to inspect to get a free
bit.

Add unit and performance tests to ensure it works as expected.

Link: https://github.com/rmind/npf/blob/99f0756f5c7a/src/kern/npf_portmap.c#L454
Signed-off-by: Robin Jarry <rjarry@redhat.com>