policy: add stateful dynamic source nat support
Add support for "stateful" dynamic source NAT. It is a convoluted term
to represent the following:

                      public server (8.8.8.8)
                    ^                        |
  source changed    |                        | conntrack lookup
     8.8.8.1        |                        | destination changed
  + ephemeral port  |                        | to 10.16.0.99
                    |                        v + original port
                +----------------------------------+
                |       iface wan   8.8.8.1/24     |
                |                                  |
                |              grout               |
                |                                  |
                |       iface lan   10.16.0.1/24   |
                +----------------------------------+
                     |                       |
                     |                       |
                     |                       v
                        client 10.16.0.99

Basically, SNAT policies contain an output interface, a private subnet
and a public IP address to be substituted for outgoing traffic. When
a SNAT policy is configured, the GR_IFACE_F_SNAT_DYNAMIC flag is added
to the wan interface. The public IP address *must* be configured prior
to creating SNAT policies.

Every time a packet originating from a client on the lan interface side
is outgoing the wan interface (one with GR_IFACE_F_SNAT_DYNAMIC), a new
connection will be created in the conntrack module (if none exists
already). The connection will be referenced in a hash table with its
forward and reverse keys (taking the translated traffic into account).

IMPORTANT: The connections and SNAT translations *must* be created
synchronously in the datapath. With an "upcall" or "exception" mechanism
that sends packets to the control plane for conntrack creation, there is
a risk that while the connection is created and packets are in queue to
be processed by the control plane thread, other packets of the same
connection could overtake the former ones, breaking packet order
preservation.

Special attention is taken to limit the interaction between multiple
datapath threads and also with the control plane thread. The NAT
translation port uses gr_id_pool which is based on atomic operations.
Pseudo random ports are reserved and allow limiting the risk of multiple
threads trying to reserve ports on the same 64 ports range. Also, a RW
lock is used to protect the SNAT policy list. Only the control plane is
able to acquire a write lock when modifying the list. Dataplane workers
only acquire read locks and will not block each others when iterating
over the list to find a matching policy when creating new connections.

The source address is replaced with the one specified in the policy. The
source port (TCP/UDP) or icmp_ident is replaced with a unique value
allocated on the fly for this specific connection.

When the public server responds to our public IP, the created connection
is looked up and used to perform the reverse translation. The connection
state is maintained for each packet.

Every second, all active connections are inspected and depending on
their state and "last-seen" activity, they may be destroyed depending on
configured timeout values.

Signed-off-by: Robin Jarry <rjarry@redhat.com>