frr: disable sanitizer
For some obscure reason, FRR headers have __attribute__((no_sanitize))
on some of their symbols. When building with -fsanitize=address, this
causes compilation errors (with GCC at least):

../frr/zebra_dplane_grout.c: In function ‘zd_grout_start_sync’:
../subprojects/frr/lib/xref.h:244:29: error: ‘no_sanitize’ attribute ignored [-Werror=attributes]
  244 |         static const struct xref * const NAMECTR(xref_p_)                      \
      |                             ^~~~

There seems not to be any way to make this work without disabling the
address sanitizer for the FRR plugin.

Remove the forced gnu11 standard for compiling the plugin, it should
inherit from the parent.

Signed-off-by: Robin Jarry <rjarry@redhat.com>