api: change request handlers to take a request context
In preparation for streaming api responses, pass a context object to all
api handlers. This argument replaces the response pointer that was to be
updated by handlers when a payload is to be attached to the response.

Instead, change the api_out structure to carry an extra pointer
containing the response payload if any.

For not the api_ctx object is only used by subscribe/unsubscribe. It
will be extended in the next commit to allow sending multiple responses
from the same api handler.

Signed-off-by: Robin Jarry <rjarry@redhat.com>