api: add support for streaming responses
Allow API handlers to return more than one message via a new api_send()
function that can be invoked with the api_ctx object, a buffer and
a size.

When API handlers use this model, they MUST terminate their stream of
responses with a final one with a zero-length payload. All responses
will carry the same for_id value matching the original request id.

Add a convenience gr_api_client_stream_foreach macro that will send the
request and iterate over all responses that match the request ID and
stop on the first error or on the first empty response.

API requests that return a stream of responses MUST be marked under
comments with the following syntax:

	// STREAM(<payload type>);

No handlers are modified yet. They will be so in multiple commits.

Signed-off-by: Robin Jarry <rjarry@redhat.com>