id-pool: make unit-tests more robust
Due to the concurrent nature of the bench tests, the steps where we
verify it works when pools are almost full sometimes result in a few
calls to gr_id_pool_get or gr_id_pool_get_random returning ENOSPC. This
is not a "bug" but a normal situation when working on highly constrained
ID pools.

Do not fail the tests when that happens. Instead, record how many ENOSPC
errors occurred and ensure that their amount does not exceed 0.01% of
the total amount of calls.

Also, after each bench test, verify the consistency of the ID pool
internal structures. Ensure that gr_id_pool_used() properly reflects the
amount of bits set to 0 and that the level0 array slabs are in sync with
level1.

Always consider errors returned by gr_id_pool_put to be fatal.

Increase the amount of iterations to 100k to increase the probability of
finding race conditions.

Signed-off-by: Robin Jarry <rjarry@redhat.com>