cli: fix broken context initialization
Depending on the order in which files are linked in the final binary,
the interface context (modules/infra/cli/iface.c) may be loaded before
the stats context (modules/infra/cli/stats.c). When the stats context
loads, there happens to be an existing "stats" node present in the
command tree:

  show interface stats
                 ^^^^^

This causes the code in get_or_create to skip the creation of the stats
context node and use this constant str node in its place. Then, when
trying to append stats commands under this str node, we get an EINVAL
error from libecoli since the base node is not an "or" node.

Replace ec_node_find() which searches recursively for *any* node that
has the provided ID. Only consider direct children of the root node and
ensure they are of type "or".

Fixes: af3ef1185218 ("cli: rework command registration")
Signed-off-by: Robin Jarry <rjarry@redhat.com>