ipv6: use input interface for responding to neigh solicitations
When receiving a neighbor solicitation, if grout owns the requested IPv6
address, we generate a neighbor advertisement for this address. If the
address is not assigned to any interface (e.g. for local SIDs),
iface_from_id(GR_IFACE_ID_UNDEF); returns NULL and it causes
a segmentation fault in icmp6_output when trying to access that
interface vrf_id.

Use the same pattern than in IPv4 output, use the input interface as
output interface.

Signed-off-by: Robin Jarry <rjarry@redhat.com>