frr: split grout_add_del_nexthop in two functions
Keep common code in grout_add_del_nexthop and fork out in two separate
functions for add and del specific code. This reduces the amount of code
and makes it easier to follow.

Also, it prepares for the next commit where nexthop structures will
become of variable size and the gr_nh_add_req structure will need to be
dynamically allocated depending on the nexthop type.

Signed-off-by: Robin Jarry <rjarry@redhat.com>