nexthop: fix corruption of telemetry stats
When DNAT or SRV6 nexthops are deallocated, their .free() callbacks are
invoked and change their type back to GR_NH_T_L3.

This causes the nh_stats.by_type[L3] to wrap around below 0 causing an
assert error when the last L3 nexthop is deleted.

  nexthop_decref: Assertion `nh_stats.by_type[nh->type] > 0' failed.

Invoke the .free() callbacks *after* the telemetry stats have been
invoked.

Also, it is possible to change the type of a nexthop via nexthop_update.
Make sure to account for this in the telemetry stats.

Fixes: 30d48691fe3d ("nexthop: only expose statistics per type")
Signed-off-by: Robin Jarry <rjarry@redhat.com>