graph: rewrite dump to work without any port configured
rte_graph_export relies on an instantiated rte_graph object which only
exists when at least one port is configured in grout.

We already contain a list of all configured node_infos which contain
a reference to the node_id, its name and flags.

The edges configured in the static structures are not updated by
rte_node_edge_update(). Hence, node edges must be resolved using
rte_node_edge_get() since they can be updated dynamically at startup
(e.g., nodes can attach themselves dynamically to eth_input such as
ip_input, arp_input or ip6_input).

Instead of exposing the length of the dump in the response message, use
a zero terminated string. This saves an extra malloc and memcpy and the
size of the message is already included in the api message header.

Signed-off-by: Robin Jarry <rjarry@redhat.com>