stats: add overhead cycles accounting
Add new loop_cycles and n_loops fields in the worker_stats structure. We
cannot reuse total_cycles which is already used by the DPDK telemetry
cpu_usage callback and which cannot be reset to 0.

Use these new fields to expose a new "overhead" meta statistic which
isn't related to a node. This software stat holds the amount of cycles
spent excluding the time spent in graph nodes and the time spent
sleeping.

Add a new counter to measure the number of main loop iterations and
attach it as the "calls" field of the new "overhead" stat.

Store the number of packets received by "port_rx" and "control_input" in
the objs field of that "overhead" stat to get an idea of what are the
overhead cycles per packet.

Signed-off-by: Robin Jarry <rjarry@redhat.com>