stats: always expose idle and overhead stats
The "idle" meta stat is not associated to a node and always has its objs
field set to 0.

The "overhead" meta stat may have objs but is still interesting to
report even when no packets are processed to compare it to "idle".

When any of these two meta stats have their calls greater than 0,
include them in the returned stats.

Signed-off-by: Robin Jarry <rjarry@redhat.com>