cli: factorize error reporting
When using grcli in interactive mode, the error reporting routine
displays where the command was found invalid and the expected tokens, if
any.

Update the batch and one-shot modes to use the same error reporting. We
cannot use ec_editline_get_suggestions which is only available in
interactive mode. Reimplement it ourselves.

Signed-off-by: Robin Jarry <rjarry@redhat.com>