api: ignore SIGPIPE when client disconnects
When a client disconnects while grout is sending data to the API socket,
the daemon may be signaled with a SIGPIPE signal.

  NOTICE: GROUT: signal_cb: received signal SIGPIPE
  ERR: GROUT: event_cb: bufferevent error on fd=16: Broken pipe

There is no guarantee which thread will execute the signal handler. To
avoid interrupting datapath threads, completely ignore the SIGPIPE (and
SIGCHLD) signals.

The write() calls will still return an error and errno will contain
EPIPE. Handle this in event_cb() and make sure to disconnect clients
when EPIPE is received.

Signed-off-by: Robin Jarry <rjarry@redhat.com>