stats: avoid crash with node_id greater than 255
The node_to_index array is used to map node IDs to offsets in the
w_stats->stats array. If we have nodes with IDs greater than 255, this
results in invalid memory accesses and can cause crashes.

Dynamically allocate this array using the maximum node ID as reported by
the rte_graph API.

Signed-off-by: Robin Jarry <rjarry@redhat.com>