stats: fix graph topological order
rte_graph_foreach_node does not iterate in the topological graph order.
At first glance it may look like so but it does not.

Use the previously introduced topo_sort generic function to sort a list
of nodes based on an is_parent callback.

Once the nodes are properly sorted, iterate over them and initialize the
topo_order field accordingly.

Fixes: 0067e940f81c ("stats: allow sorting by graph topological order")
Signed-off-by: Robin Jarry <rjarry@redhat.com>