smoke: print stack traces on crashes
The smoke tests suite does not provide any information when grout
crashes with a signal. This makes debugging very hard, especially in CI
environments.

Capture the grout PID at startup with pgrep to be able to use it later
in the cleanup trap function. Set ulimit to allow core dumps, even when
ASAN is enabled. Configure the kernel core pattern to write core dump
files in the temporary test folder.

In the cleanup EXIT trap, save the current script exit status into
a variable to avoid it being trampled by subsequent commands. When
waiting for the grout process, inspect the returned status code to
detect if it was terminated by a signal.

If a core dump is available, use gdb to extract stack traces from all
threads. As a fallback, when no custom core pattern could be set, use
systemd-coredumpctl to gather information about the crash.

Restore the trap exit status after all cleanup has been done to
preserve the original script exit status and avoid marking failed tests
as successful.

This requires adding gdb as a dependency for running the smoke tests.

Signed-off-by: Robin Jarry <rjarry@redhat.com>