ip6: send neigh advert for every address when iface is up
When a network interface transitions to the UP state, send unsolicited
neighbor advertisements for all IPv6 addresses configured on that
interface. This allows the network to immediately update forwarding
tables with the correct MAC address, which is critical in bonding
scenarios where the active member may change.

Unsolicited neighbor advertisements are sent by calling nh6_advertise()
with the local address nexthop and NULL for the remote parameter. These
packets are sent to the all-nodes multicast address (ff02::1) and have
the solicited flag cleared.

This mirrors the gratuitous ARP functionality implemented for IPv4
addresses.

Signed-off-by: Robin Jarry <rjarry@redhat.com>