main: don't probe devices on fslmc bus
Like pci bus, we should not probe any devices on fslmc bus.
Let's add a fake device in the allowed list of DPDK.

Note: This series doesn't work as fslmc bus do nothing to plug or unplug
a device. [1]. So it should not be merged until it has been fixed in the
DPDK.

[1] https://git.dpdk.org/dpdk/tree/drivers/bus/fslmc/fslmc_bus.c#n591
Signed-off-by: Maxime Leroy <maxime@leroys.fr>